+++ packages/react/src/components/Select/styles.ts
@@ -1,17 +1,18 @@
 import type { ComponentProps } from 'react'
 import { styled } from '../../styles'
 
-export const SelectUI = styled('select', {
-  gap: '$2',
-  overflow: 'hidden',
-
-  background: 'transparent',
-  border: 'none',
+export const SelectContainer = styled('div', {
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'space-between',
 
   fontFamily: '$app',
   color: '$white',
-  fontWeight: '$regular',
+
   cursor: 'pointer',
+
+  width: '100%',
+  maxWidth: '24.5rem',
 
   boxSizing: 'border-box',
   borderRadius: '$md',
@@ -20,18 +21,80 @@ export const SelectUI = styled('select', {
 
   transition: 'border 0.2s linear',
 
-  width: '100%',
-  margin: 0,
-  padding: '$1 $1',
+  '&:disabled': {
+    cursor: 'not-allowed',
+    opacity: 0.5,
+  },
 
-  '&:focus': {
-    outline: 0,
+  '&:active': {
     borderBottom: '2px solid $dafaz400',
   },
 
+  '&:hover': {
+    borderBottom: '2px solid $dafaz400',
+  },
+
+  variants: {
+    size: {
+      lg: {
+        fontSize: '$xl',
+        svg: {
+          width: '$5',
+          height: '$5',
+        },
+      },
+      md: {
+        fontSize: '$lg',
+        svg: {
+          width: '$4',
+          height: '$4',
+        },
+      },
+      sm: {
+        fontSize: '$sm',
+        svg: {
+          width: '$3',
+          height: '$3',
+        },
+      },
+    },
+  },
+  defaultVariants: {
+    size: 'lg',
+  },
+})
+
+interface SelectContainer extends ComponentProps<typeof SelectContainer> {
+  size?: 'sm' | 'md' | 'lg'
+}
+
+export const SelectUI = styled('select', {
+  zIndex: 1,
+  '-moz-appearance': 'none',
+  '-webkit-appearance': 'none',
+  appearance: 'none',
+
+  overflow: 'hidden',
+  background: 'transparent',
+  border: 'none',
+
+  fontFamily: '$app',
+  color: '$white',
+  fontWeight: '$regular',
+  cursor: 'pointer',
+
+  width: '100%',
+  maxWidth: '24.5rem',
+
+  margin: 0,
+  padding: '$1 $1',
+
   '&:disabled': {
     opacity: 0.5,
-    cursor: 'not-allowed',
+  },
+
+  '&:focus': {
+    outline: 0,
   },
 
   variants: {